devops/docker/05-compose/code/api/requirements.txt
# Flask web framework
flask==3.1.2

# PostgreSQL adapter
psycopg2-binary==2.9.11

# Production WSGI server
gunicorn==23.0.0
